backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-environ
django-filter
openai>=1.0.0
google-genai>=1.0.0
